+++ packages/xo-web/src/common/intl/messages.js
@@ -197,6 +197,11 @@ const messages = {
   stateDisabled: 'Disabled',
   stateEnabled: 'Enabled',
 
+  // ----- Labels -----
+  labelMerge: 'Merge',
+  labelSize: 'Size',
+  labelTransfer: 'Transfer',
+
   // ----- Forms -----
   formCancel: 'Cancel',
   formCreate: 'Create',
@@ -349,6 +354,7 @@ const messages = {
   missingSnapshotRetention:
     'The Rolling Snapshot mode requires snapshot retention to be higher than 0!',
   retentionNeeded: 'One of the retentions needs to be higher than 0!',
+  newScheduleError: 'Invalid schedule',
   createRemoteMessage:
     'No remotes found, please click on the remotes settings button to create one!',
   remotesSettings: 'Remotes settings',
@@ -854,6 +860,8 @@ const messages = {
   started: 'Started {ago}',
   paraVirtualizedMode: 'Paravirtualization (PV)',
   hardwareVirtualizedMode: 'Hardware virtualization (HVM)',
+  hvmModeWithPvDriversEnabled:
+    'Hardware virtualization with paravirtualization drivers enabled (PVHVM)',
   windowsUpdateTools: 'Windows Update tools',
 
   // ----- VM stat tab -----